refactor(rust-sdk): centralize unsupported network backend
