+++ src/boxlite/src/runtime/rt_impl.rs
@@ -25,7 +25,7 @@ fn litebox_from_impl(box_impl: SharedBoxImpl) -> LiteBox {
     let network_backend: Arc<dyn crate::runtime::backend::BoxNetworkBackend> = box_impl.clone();
     let snapshot_backend: Arc<dyn crate::runtime::backend::SnapshotBackend> =
         Arc::new(LocalSnapshotBackend::new(box_impl));
-    LiteBox::new_with_network(box_backend, network_backend, snapshot_backend)
+    LiteBox::new(box_backend, network_backend, snapshot_backend)
 }
 
 /// Archive the active exit file as `exit.previous`, freeing the canonical
